% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELOS_mat_4CE.R
\name{ELOS_mat}
\alias{ELOS_mat}
\title{Expected Length of Stay Matrix}
\usage{
ELOS_mat(pt, end_day, sub_condition, boot_n)
}
\arguments{
\item{pt}{probability transition matrix}

\item{end_day}{days after admission for which estimation should occur}

\item{sub_condition}{subgroup to be analyzed}

\item{boot_n}{number of bootstrap samples for confidence inteval estimation}
}
\value{
list with expected length of stay matrix and bootstrapped confidence intervals
}
\description{
Function for creating expected length of stay and probability of being in hospital matrix
}
