% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_mstate_4CE.R
\name{ext_mstate}
\alias{ext_mstate}
\title{Extend data frame for mstate}
\usage{
ext_mstate(dataset, traM)
}
\arguments{
\item{dataset}{data frame of observed transitions}

\item{traM:}{transition matrix of multi-state model used in mstate}
}
\value{
data frame with rows of all possible transitions for which a patient is at risk
}
\description{
Prepares data frame for use in mstate package
}
