% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOS_boot_4CE.R
\name{LOS_boot}
\alias{LOS_boot}
\title{Bootstrap Length of Stay}
\usage{
LOS_boot(data1, days_from_adm)
}
\arguments{
\item{data1}{data frame with rows of all possible transitions for which a patient is at risk (i.e. returned data frame from ext_mstate function)}

\item{days_from_adm}{number of days since admission for which estimation of duration in states is performed}
}
\value{
vector with c(1,2,...): non-severe/severe duration estimates for patients non-severe at admission   c(...,3,4,...): non-severe/severe duration estimates for patients severe at admission   c(...,5): length of severe stay estimates for entire cohort
}
\description{
Function for bootstrap confidence intervals
}
