% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_V21.R
\name{prep_data}
\alias{prep_data}
\title{Prepare Data}
\usage{
prep_data(dir.input, last_day)
}
\arguments{
\item{dir.input}{Path to the input file}

\item{last_day}{Last date when the local data set was updated}
}
\value{
The data frame in long format with covariates
}
\description{
This function creates a data frame in long format with the covariates of interest
}
