% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_CR.R
\name{prep_data_CR}
\alias{prep_data_CR}
\title{Prepare Competing Risks}
\usage{
prep_data_CR(data_set)
}
\arguments{
\item{data_set}{data frame to be converted}
}
\value{
data frame prepared for competing risks model
}
\description{
Create data set for competing risks model
}
