% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_tran_mat_4CE.R
\name{prob.tran.mat}
\alias{prob.tran.mat}
\title{Creates probability transition matrix}
\usage{
prob.tran.mat(data1, sub_condition, trans_matrix)
}
\arguments{
\item{sub_condition}{subset of data to be analyzed}
}
\value{
transition matrix, cumulative hazard increments, and cox model results
}
\description{
Creates and saves probability transition matrix, cumulative hazard increments, and cox model results in list
}
