% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_4CE.R
\name{run_analysis}
\alias{run_analysis}
\title{Run analysis}
\usage{
run_analysis(dir_LocalPatientSummary, dir.output, site_name, adm_cens, workd1)
}
\arguments{
\item{dir_LocalPatientSummary}{path to the Local Patient Summary .csv file}

\item{dir.output}{directory to results}

\item{site_name}{name of local site}

\item{adm_cens}{last day of data collection}
}
\value{
results in html and R object
}
\description{
runs multi-state analyses
}
