% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{Save results}
\usage{
save_results()
}
\value{
list with cumulative hazards, expected lengths of stay and Cox model summaries
}
\description{
Saves aggregated data as list
}
